gptel: Use FSM to drive gptel requests
* gptel.el (gptel-request, gptel-abort, gptel-send,
gptel--url-get-response, gptel--url-parse-response,
gptel--insert-response): Plug `gptel-request' into the state
machine.  This makes `gptel-request' follow the control flow
specified by its `fsm' argument, which should be a `gptel-fsm'
struct.  In addition to making the flow configurable, this cleanly
separates the parts of the request process that are specific to
`gptel-send', such as updating the UI etc.  We can now turn them
off/on as required easily.  Make `gptel--url-get-response' work
correctly when making a second or later request with the same
FSM.  (This is required for tool calls.)

* gptel-curl.el (gptel-curl--parse-response, gptel-curl--sentinel,
gptel-curl--stream-filter, gptel-curl--stream-cleanup,
gptel-curl-get-response): Simplify the various process sentinels and
filters by (i) using the fsm, calling state transitions where
appropriate, and (ii) by removing all the UI update code.  Make
`gptel-curl-get-response' work correctly when making a second or
later request with the same FSM.