gptel: Add a diagnostic display for latest request
* gptel.el (gptel-request--handlers, gptel--fsm-last,
gptel--fsm-inspect, gptel--update-status): Add function and
variable `gptel--fsm-last' to capture the state of the latest
request.  The function `gptel--fsm-inspect' can be used to inspect
(most of) it in a separate buffer.  Clicking on the "Ready" (etc)
status text in the header-line in gptel chat buffers now calls
`gptel--fsm-inspect'.

The intent of this function is to make introspection more
convenient.